[Refactoring] Separated data struct to each files and cleaning codes
