+++ AmigoPJT/AmigoPJT.vcxproj
@@ -141,18 +141,9 @@
   <ItemGroup>
     <ClCompile Include="amigoDatabase.cpp" />
     <ClCompile Include="client.cpp" />
-    <ClCompile Include="client.h" />
     <ClCompile Include="InputStream.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="outputStream.cpp" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="amigoDatabase.h" />
-    <ClInclude Include="common.h" />
-    <ClInclude Include="iDatabase.h" />
-    <ClInclude Include="inputStream.h" />
-    <ClInclude Include="search.h" />
-    <ClInclude Include="outputStream.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -160,6 +151,18 @@
   <ItemGroup>
     <Text Include="input_20_20.txt" />
     <Text Include="output_20_20.txt" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="amigoDatabase.h" />
+    <ClInclude Include="client.h" />
+    <ClInclude Include="command.h" />
+    <ClInclude Include="common.h" />
+    <ClInclude Include="employee.h" />
+    <ClInclude Include="iDatabase.h" />
+    <ClInclude Include="inputStream.h" />
+    <ClInclude Include="outputStream.h" />
+    <ClInclude Include="search.h" />
+    <ClInclude Include="util.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
