+++ AmigoPJT/AmigoPJT.vcxproj.filters
@@ -33,9 +33,6 @@
     <ClCompile Include="$(MSBuildThisFileDirectory)..\..\lib\native\src\gmock\gmock-all.cc">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="client.h">
-      <Filter>헤더 파일</Filter>
-    </ClCompile>
     <ClCompile Include="amigoDatabase.cpp">
       <Filter>소스 파일\Database</Filter>
     </ClCompile>
@@ -58,7 +55,25 @@
     </Text>
   </ItemGroup>
   <ItemGroup>
+    <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="amigoDatabase.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="client.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="command.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
     <ClInclude Include="common.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="employee.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="iDatabase.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="inputStream.h">
@@ -70,14 +85,8 @@
     <ClInclude Include="search.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
-    <ClInclude Include="amigoDatabase.h">
+    <ClInclude Include="util.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
-    <ClInclude Include="iDatabase.h">
-      <Filter>헤더 파일</Filter>
-    </ClInclude>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="packages.config" />
   </ItemGroup>
 </Project>
