+++ AmigoPJT/iDatabase.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <string>
-#include "common.h"
+#include "command.h"
 
 class IDatabase
 {
