+++ AmigoPJT/outputStream.h
@@ -3,7 +3,8 @@
 #include <fstream>
 #include <iostream>
 #include <string>
-#include "common.h"
+
+using namespace std;
 
 class OutputStream
 {
